SAP: Fix crash while closing SAP Service
Usecase:
1) Perorm BT off and on stress testing.

Expectation:
BT should be switched off and on without issues

Observed Result:
BT Off failed in some iteration due to crash in SapService

Rootcause:
In some rare cases, when ProfileService calls stop and cleanup in
quick succession, mSessionStatusHandler becomes null while call to
closeService() from stop(), so handling of sendShutdownMessage leads
to crash as the Service is already stopped and resources freed.

Fix:
Add a flag to track is SapService is already stopped or in process
of being stopped to prevent multiple call to closeService.

CRs-Fixed: 2075124
Change-Id: I0968fd6569dedfe0c11449821642ac91e6ab4c99