AVRCP: Avoid trailing whitespace for title attribute.
Some players add or remove trailing or leading whitespaces
to title field each time when screen off/on is done.
This causes avrcp to treat it as track change and send
update to remote causing some remotes to behave bad.
This fix removes any leading or trailing whitespaces in
title field.

CRs-Fixed: 2056886
Change-Id: If0a280afcb03b0ac4e9c39af456eedf1eba0d27c