AVRCP: Delay Media Session Registration when HF is in call
Problem: Meta data does not update on carkit after VoIP call.

Cause: Registration for Media session callback occurs as soon
as the call is disconnected. At that time, HF is yet to get
sco termination call. So, media session registration is ignored
because HF reports an active call.

Fix: Added delay to the processing of set media session call
when HF call is active instead of ignoring it.

Change-Id: Id30d90335b2a4ee73d74a6c1bfb947628cd5aba1
CRs-Fixed: 2152756