AVRCP: Assign proper playstate when avrcp connects.
When AVRCP connects, it checks for current player state
saved locally(mCurrentPlayerState) and a2dp playing state.
If current player state saved locally is NOT PLAYING but a2dp
state is streaming due to either touch tones or remote start,
it used to update current playstate for this device as PLAYING.
This causes PLAYING playback state incorrectly being update to
the remote device. The code change adds isMusicActive() check
to ensure play state is not updated incorrectly in such case.

Change-Id: I023234fd88cab96ee769847a56e3e024c6ad7828