Handle passthrough commands which are repeated from remote.
There are some players taking time to execute passthrough
commands, in this case if remote will initiate passthrough
which are different from previous passthrough then it is
not excuted at all. So to take care of these type of scenarios
cache the last passthrough commands, when it differs with
current command, then dispatch the current command too to
take corresponding action.

Change-Id: I3002a1078d6da8ee766a0d03de7725e8e68c5c77