Adding a new BroadcastReceiver for phone state change
When the SIM state is changed i.e loaded or unloaded,
the simStateChanged intent is received.
The  BroadcastReceiver mPhoneStateChangeReceiver
is registered during bluetooth ON and unregistered
during bluetooth OFF.
This BroadcastReceiver listens to the simstatechange
always when BT is ON. Signal strength and service
availability values sent to connected devices depending
on the sim state.

CRs-Fixed: 2202981
Change-Id: I4f903afd07e2719cee2ee093edfdc5526fd865e4