HFP: Query the phone states during multiHFP scenario
  During the dual HFP scenario, the first headset
with CVSD codec is connected. During the call, the
SCO is established and the call audio is heard on the
first headset. Now when the second headset is connected
with mSBC codec, the MM audio incorrectly selects
second headset even though SCO is not connected.
This results in audio lose.

  It is addressed by querying for phone state after
SLC done with second headset i.e. QUERY_PHONE_STATE_AT_SLC
is handled in AudioOn state.

CRs-Fixed: 2182377
Change-Id: I5cef3f76bcb3eec498bd918dcd61d2219c4a13f6