Bluetooth-A2DP: Set volume of A2DP SNK stream back to the previous
The scenario is making an outgoing call while A2DP SNK is streaming. When the call is hang up,
the volume of resumed A2DP SNK stream is very low so that user can hardly hear the sound.

The workaround is to set volume of A2DP SNK stream back to the value paused when A2DP SNK
streaming is resumed.

Change-Id: I4855312cba7aaf481662cc7a124b847784f9decf
Signed-off-by: Yanfei Zhou <zyanfei@codeaurora.org>