Avrcp: Sending available player CHANGED response during registration
Earlier when BT turns ON then mAvailablePlayerviewchanged flag is set true,
and because of this when we send A2DP START or AVRCP play from DUT,
then in first attempt CHANGED response gets ignore and we retured
without updating current play status to remote.

Now we have shifted this logic to available player changed event,
where this flag will be set as false during registration.

CRs-Fixed: 2238114
Change-Id: Ic158af7bf0d58ad0aa8557a2840da6d02583ed4c