A2DP Sink: Broadcast Connection State when Bluetooth Process is killed
Usecase:
1) Establish connection with DUT in A2DP Sink Role.
2) Go to Setting->Network and Internet Setting and Reset.
3) Bluetooth is turned off and turned on with all Bonded
   devices removed from list.

Issue:
Bluetooth connection state is not updated on UI in Setting
menu or Notification bar when network reset is done or
Bluetooth Process is killed.

Root Cause:
As disconnected state was not broadcasted, Setting App was
not reflecting change in Bluetooth Connection State Icon.

Fix:
Broadcast connection state as DISCONNECTED when Service is
destroyed or stopped.

Change-Id: Ic769a61ad08a682cfc60ce953dc9500aa4bd2663
CRs-Fixed: 2234061