Bluetooth: Support to update call id
Currently, two hash tables mCalls and mCallsUpdate are created in HeadsetClientStateMachine
to maintain status of all calls. If an outgoing call is initiated from HF, call id will be
set as (-1) when the object of this call is created. It is dedicated to design as this.
The key (id = -1) of the outgoing call in hash table will be associated with one of the
mCallsUpdated calls, the field (mId) in the object will not be updated.

This change is to update the call id in the call object.

Change-Id: I070fe40c8266131b26cd06f961bd843c970c7070
Signed-off-by: Song Gao <songg@codeaurora.org>