Send call indicators to remote in Serialized way.
When MO call created and disconnected back to back, we are sending the call
disonnected indicator first, and then Call dialling and Call alerting, which
will keep remote into bad state and A2dp didn't resumed. So send the call
indicators in serialized way as +CIEV 2,2., +CIEV 2,3., and +CIEV 2,0.

CRs-Fixed: 2289855
Change-Id: If601851dc43d84c15f69acd9451d21fa75b2588a