HFP: Add new variable to maintain call UUID
Currently, in the implementation of equals() in HfpClientConnection.java,
it calls getUUID(). In some use cases such as multi-party conference call,
when getUUID() is called, the member variable mCurrentCall is already
invalid.

This code change is to add a new member variable "mUUID" to maintain and
update the value of "mCurrentCall.getUUID()".

Change-Id: Ib182da51227f859ddf85eb22f49078f88b8d8b6b
Signed-off-by: Song Gao <songg@codeaurora.org>