PBAP_Client: Check handler before sending message
Test:
1. Pair with remote device
2. Turn off bluetooth on remote device
3. Turn off bluetooth on DUT
4. Turn on bluetooth on DUT
5. During connect pbap turn off bluetooth on DUT

Result:
A null pointer reference can be found

Root Cause:
When BT is turned off, PbapClientService is stopped
so as to make PbapClientStateMachine quit, while
PbapClientConnectionHandler may be blocked in
establishing socket or obext session.
When PbapClientStateMachine quit, SmHandler is cleaned up.
So NullPointerException is raised in PbapClientConnectionHandler
when to send message to PbapClientStateMachine.

Fix:
Check smHandler to avoid NullPointerException when to sendToTarget

Change-Id: I0f7a98c3f38a06884d74f0605fd7de1e88f42d4b
Signed-off-by: Fei Zheng <zfei@codeaurora.org>