AVRCP_CT: Use listener to update folder list items
In current implementation, intent is used to share updating folder list
items with A2dpMediaBrowserService. However, it is an ugly design for
failed to cover the use case of large size browsing list. VM will carsh
for bundle exception.

This code change abandon the intent implementation, use listener to
listen the update of folder list items.

Change-Id: I11880a63549297d09e119d689b8b55b63ef34628
Signed-off-by: Song Gao <songg@codeaurora.org>