AVRCP_BROWSE: Avoid using iterator for large size list operation
Currently, in the implementation of msgFolderList(), foreach is used to copy list
from List<Parcelable> to List<MediaItem>. foreach in list is implemented by
iterator, which causes ConcurrentModificationException if the source list is very
large in size.

This code change is to avoid using iterator for large size list operation.

Change-Id: I3d07b95fc17c21c3fcfca5274078e0fb0d020b76