AVRCP: Change AVRCP_ENABLE_PROPERTY to fix selinux issue
Reason:
(1) Originally, AVRCP_ENABLE_PROPERTY is set as
'persist.bluetooth.enablenewavrcp', which is
defined as bluetooth_prop.

(2) If selinux is set as enforcing, the property
can't be set by vendor_init so that AVRCP_CT
can't be enabled for Automotive.

Resolution:
Change AVRCP_ENABLE_PROPERTY as 'persist.vendor.bluetooth.enablenewavrcp',
which is defined as vendor_bluetooth_prop

Change-Id: Id210f223af83b6ab8fb3807e14c3f12d82d47ae2