HFP_HF: Fix deadlock issue during HFP call
Description:
Car dialer blocks in calling getCurrentCalls which
finnally leads to ANR.

Root Cause:
When the SCO is connected, the thread will hold the
synchronized lock in class HeadsetClientStateMachine,
and try to get synchronized lock in class
HeadsetClientService.
E.g.
1. Synchronized function processMessage in State
Connected is called.
2. Accordingly, mixHfpAudio will be called
3. It will try to call getConnectedDevices in HeadsetClientService
But may be held till if it can't get synchronized lock in
HeadsetClientService.

In the meantime, APPs like Car dialer
may call AIDL APIs. So this thread may hold the synchronized
lock in HeadsetClientService at first, then tries to get
synchronized lock in HeadsetClientStateMachine.
E.g.
1. Car dialer calls API getCurrentCalls.
2. Accordingly, synchronized function getCurrentCalls in
HeadsetClientService is called.
3. It will try to call getConnectionState in HeadsetClientStateMachine.
But may be held if it can't get synchronized lock in
HeadsetClientStateMachine.

Solution:
Avoid calling synchronized functions in HeadsetClientService
from HeadsetClientStateMachine when current thread already
owns the synchronized lock of HeadsetClientStateMachine.

Note:
Only one SCO is allowed if dual SCO feature is not enabled.
So it shall be good to keep calling of HeadsetClientService.isScoRouted
from HeadsetClientStateMachine.

Change-Id: I8123eca4d0bfd477700d454a2e8c5f58c6d91d54