AVRCP_CT: Set browsing boundary value according to item count in folder.
Issue description:

CT is able to get a maximum of 1000 item when performing
"GET FOLDER ITEMS", the other items are ignored.

This boundary is limited by MAX_FOLDER_ITEMS (1000)

Solution:

When the items count in VFS is retrieved, set this value
as browsing scope's boundary, rather than the smaller
one between this value and MAX_FOLDER_ITEMS (1000).

Change-Id: Iee207262d5b9671e2bab4efd811aaae6a5f8e6dd